Address remaining review comments & run prettier
