+++ valohai_yaml/schema/task.yaml
@@ -1,4 +1,3 @@
----
 type: object
 required:
   - name
@@ -21,7 +20,12 @@ properties:
     type: string
   type:
     type: string
-    enum: [ "bayesian_tpe", "grid_search", "distributed", "random_search", "manual_search" ]
+    enum:
+      - bayesian_tpe
+      - distributed
+      - grid_search
+      - manual_search
+      - random_search
   name:
     type: string
     description: The unique name for this task.
