+++ valohai_yaml/schema/variant-param-rule-item.yaml
@@ -1,4 +1,3 @@
----
 type: object
 additionalProperties: false
 properties:
@@ -19,7 +18,7 @@ properties:
   items:
     type: array
     items:
-        type: number
+      type: number
   min:
     type: number
   max:
@@ -29,4 +28,4 @@ properties:
   seed:
     type: number
   integerify:
-      type: boolean
+    type: boolean
